test(proxy): cover Mistral passthrough in CI shard
